Move run_single into runnables
